requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
psycopg2-binary>=2.9.11
python-telegram-bot>=20.7
python-dotenv>=1.0.0
lxml>=4.9.3
fake-useragent>=1.4.0
schedule>=1.2.0


